refactor(admin): move AdminError to dedicated error.rs
